NXP-20452: extract 3D content metadata; add to main file and transmission formats; display information on JSF

+++ nuxeo-platform-3d-jsf/src/main/java/org/nuxeo/ecm/platform/threed/transmissionFormat/TransmissionFormatItem.java
@@ -114,7 +114,7 @@ public class TransmissionFormatItem {
     }
 
     public String getTextureSize() {
-        Float texturesSize = info.getTexturesSize();
+        Long texturesSize = info.getTexturesSize();
         return (texturesSize == null) ? "-" : texturesSize.toString();
     }
